+++ frontend/src/constants.ts
@@ -1,0 +1,1 @@
+export const USER_CLOSE_WEB_SOCKET_CODE = 4333;
