+++ frontend/src/types.ts
@@ -10,5 +10,3 @@ export enum AppState {
   CODING = "CODING",
   CODE_READY = "CODE_READY",
 }
-
-export const USER_CLOSE_WEB_SOCKET_CODE = 4333;
